requirements.txt
aiohttp>=3.8.0
jinja2>=3.1.0
pyyaml>=6.0.0
typer>=0.9.0
rich>=13.0.0
